experiments/requirements.txt
# experiments/requirements.txt

mlflow
scikit-learn
prometheus_client
python-dotenv
boto3
joblib